requirements.txt
requests
python-dotenv
openai
python-telegram-bot
